clean up code and lint errors
