+++ src/components/IconBar/IconBar.tsx
@@ -1,13 +1,8 @@
-/* eslint-disable @typescript-eslint/indent */
+
 import React from 'react';
 import PropTypes from 'prop-types';
 import { makeStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
-// import GridList from '@material-ui/core/GridList';
-// import GridListTile from '@material-ui/core/GridListTile';
-import clsx from 'clsx';
-
-import { DeviceContext } from 'contexts/Device';
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -39,7 +34,6 @@ interface Props {
 
 export default function IconBar({ children, justify }: Props) {
     const classes = useStyles();
-    const value = React.useContext(DeviceContext);
 
     return (
         <div className={classes.root}>
@@ -54,9 +48,6 @@ export default function IconBar({ children, justify }: Props) {
                         {React.Children.map(children, (child, idx) => {
                             // clone the child and apply the button styles
                             const StylizedChild = React.cloneElement(child, {
-                                className: clsx([
-                               
-                                ]),
                                 size: 'large',
                                 variant: 'contained',
                             });
@@ -74,28 +65,6 @@ export default function IconBar({ children, justify }: Props) {
                     </Grid>
                 </div>
             </div>
-
-            {/* <GridList cellHeight={160} className={classes.gridList} cols={12}>
-                {React.Children.map(children, (child, idx) => {
-                    // clone the child and apply the button styles
-                    const StylizedChild = React.cloneElement(child, {
-                        className: clsx([classes.btn, child.props.className]),
-                        size: 'large',
-                        variant: 'contained',
-                        color: 'primary',
-                        elevation: 5,
-                    });
-
-                    return (
-                        <GridListTile
-                            key={idx}
-                            cols={btnCols || sizing[value] || 3}
-                        >
-                            {StylizedChild}
-                        </GridListTile>
-                    );
-                })}
-            </GridList> */}
         </div>
     );
 }
